+++ test/fixtures/binding.js
@@ -1,0 +1,4 @@
+const VrpcFactory = require('../../vrpc/VrpcFactory')
+const TestClass = require('./TestClass')
+
+VrpcFactory.register(TestClass)